型定義変更

+++ db/CreateTables.sql
@@ -120,12 +120,12 @@ CREATE TABLE MAKERS (
 
 --機器マスタ
 CREATE TABLE EQUIPMENTS (
-   id          int          NOT NULL identity
-  ,name        nvarchar(20) NOT NULL
-  ,rating      float        NULL
-  ,maker_id    int          NOT NULL
-  ,enable_flag bit          NOT NULL
-  ,upd_date    datetime     NOT NULL
+   id          int           NOT NULL identity
+  ,name        nvarchar(20)  NOT NULL
+  ,rating      decimal(5, 2) NULL
+  ,maker_id    int           NOT NULL
+  ,enable_flag bit           NOT NULL
+  ,upd_date    datetime      NOT NULL
   ,primary key(id)
 )
